requirements.txt
APScheduler==3.11.2
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.3.1
django-crontab==0.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==25.0
pillow==12.1.0
praytimes==2.3.2
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
setuptools==80.9.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
wheel==0.45.1
whitenoise==6.6.0
